GPL_RT_AX86U_300438641535

+++ release/src/router/www/Advanced_IPTV_Content.asp
@@ -60,21 +60,45 @@ var orig_wan_vpndhcp = '<% nvram_get("wan_vpndhcp"); %>';
 var orig_ttl_inc_enable = '<% nvram_get("ttl_inc_enable"); %>';
 var iptv_profiles = [<% get_iptvSettings();%>][0];
 var isp_profiles = iptv_profiles.isp_profiles;
+var orig_wnaports_bond = '<% nvram_get("wanports_bond"); %>';
+
 if(wan_bonding_support)
 	var orig_bond_wan = httpApi.nvramGet(["bond_wan"], true).bond_wan;
+
+// get Primary WAN setting list from hook.
+// cplumn 0: dsl(wan)_enable
+var MSWAN_List_Pri = [ <% get_MS_WAN_list_Pri(); %> ];
+var mr_mswan_idx_orig = '<% nvram_get("mr_mswan_idx"); %>';
+var with_multiservice=0;
+for(var i = 1; i < MSWAN_List_Pri.length; i++){
+	if (MSWAN_List_Pri[i][0] == "1") {
+		with_multiservice++;
+	}
+}
 
 function initial(){
 	show_menu();
 
+	create_stb_select(original_switch_stb_x);
+	if(mswan_support){
+		update_mr_mswan_idx();
+		inputCtrl(document.form.mr_mswan_idx, with_multiservice ? 1 : 0);
+	}
+	create_mr_select(orig_mr_enable);
 	if(dsl_support) {
-		document.form.action_script.value = "reboot";
-		document.form.action_wait.value = "<% get_default_reboot_time(); %>";
+		document.form.action_wait.value = 20;
+		if( based_modelid == "DSL-AC68U"){
+			$("#dsl_vlan_check").show();
+			document.form.action_script.value = "reboot";
+			document.form.action_wait.value = "<% get_default_reboot_time(); %>";
+		}
+		$("#isp_profile_tr").hide();
+		$("#mr_enable_field").show();
+		$("#enable_eff_multicast_forward").show();
 	}
 	else{	//DSL not support
-		create_stb_select();
 		create_ISP_select();
 		ISP_Profile_Selection(original_switch_wantag);
-		document.form.switch_stb_x.disabled = false;
 	}
 	
 	if(vdsl_support) {
@@ -102,10 +126,6 @@ function initial(){
 		document.getElementById("port_settings").style.display = "";
 		document.form.iptv_port_settings.disabled = false;
 		change_port_settings(iptv_port_settings_orig);
-	}
-
-	if(hnd_support){
-		$('#igmp_default_ver_field').show();
 	}
 }
 
@@ -250,10 +270,8 @@ function ISP_Profile_Selection(isp){
 
 	iptv_modified = 0;
 	voip_modified = 0;
-	if(isp == "none"){
+	if(isp == "none")
 		document.getElementById("wan_stb_x").style.display = "";
-		create_stb_select(isp_settings.switch_stb_x);
-	}
 	else
 		document.getElementById("wan_stb_x").style.display = "none";
 
@@ -337,12 +355,15 @@ function ISP_Profile_Selection(isp){
 	if(isp_settings.mr_enable_x != ""){
 		document.getElementById("mr_enable_field").style.display = "none";
 		document.form.mr_enable_x.disabled = true;
+		change_mr_enable(0);
 	}
 	else{
 		document.getElementById("mr_enable_field").style.display = "";
 		document.form.mr_enable_x.disabled = false;
 		document.form.mr_enable_x.value = orig_mr_enable;
-		change_mr_enable(isp_settings.switch_stb_x); //hnd or BLUECAVE
+		if(document.form.mr_enable_x.selectedIndex < 0)
+			document.form.mr_enable_x.selectedIndex = 0;
+		change_switch_stb(isp_settings.switch_stb_x); //hnd or BLUECAVE
 	}
 
 	if(isp_settings.emf_enable != ""){
@@ -367,6 +388,7 @@ function ISP_Profile_Selection(isp){
 function validForm(){
 	if (!dsl_support){
         if(document.form.switch_wantag.value == "manual"){
+			document.form.switch_stb_x.disabled = false;
 			if(document.form.switch_wan1tagid.value == "" && document.form.switch_wan2tagid.value == "")
 				document.form.switch_stb_x.value = "0";
 			else if(document.form.switch_wan1tagid.value == "" && document.form.switch_wan2tagid.value != "")
@@ -395,14 +417,19 @@ function validForm(){
 				return false;
 		}
 		else if(document.form.switch_wantag.value == "none"){
+			document.form.switch_stb_x.disabled = false;
 			document.form.switch_stb_x.value = document.form.switch_stb_x0.value;
 		}
+		else{
+			var isp_profile = get_isp_settings(document.form.switch_wantag.value);
+			document.form.switch_stb_x.value = isp_profile.switch_stb_x;
+		}
+	}
+	else{
+		document.form.switch_stb_x.disabled = false;
+		document.form.switch_stb_x.value = document.form.switch_stb_x0.value;
 	}
 
-	return true;
-}
-
-function applyRule(){
 	if(dualWAN_support){	// dualwan LAN port should not be equal to IPTV port
 		var tmp_pri_if = wans_dualwan_orig.split(" ")[0].toUpperCase();
 		var tmp_sec_if = wans_dualwan_orig.split(" ")[1].toUpperCase();
@@ -413,7 +440,7 @@ function applyRule(){
 				port_conflict = true;
 			else if( (wans_lanport == 1 || wans_lanport == 2) && iptv_port == 5)
 				port_conflict = true;
-			else if( (wans_lanport == 3 || wans_lanport == 4) && iptv_port == 6)
+			else if( (wans_lanport == 3 || wans_lanport == 4) && (iptv_port == 6 || iptv_port == 8) )
 				port_conflict = true;
 
 			if (port_conflict) {
@@ -423,6 +450,23 @@ function applyRule(){
 		}
 	}
 
+	return true;
+}
+
+function turn_off_lacp_if_conflicts(){
+	if (!lacp_enabled)
+		return;
+
+	if((based_modelid == "RT-AX89U" || based_modelid == "GT-AXY16000")){
+		// LAN1 and/or LAN2.
+		if(document.form.switch_wantag.value == "none" && (document.form.switch_stb_x0.value == 1 || document.form.switch_stb_x0.value == 2 || document.form.switch_stb_x0.value == 5)){
+			document.form.lacp_enabled.disabled = false;
+			document.form.lacp_enabled.value = "0";
+		}
+	}
+}
+
+function applyRule(){
 	if(!dsl_support){
 		if( (original_switch_stb_x != document.form.switch_stb_x0.value)
 			|| (original_switch_wantag != document.form.switch_wantag.value)
@@ -433,6 +477,7 @@ function applyRule(){
 																	|| (original_switch_wan2tagid != document.form.switch_wan2tagid.value)
 																	|| (original_switch_wan2prio != document.form.switch_wan2prio.value)) )
 			){
+			turn_off_lacp_if_conflicts();
 			FormActions("start_apply.htm", "apply", "reboot", "<% get_default_reboot_time(); %>");
 		}
 	}
@@ -482,8 +527,8 @@ function applyRule(){
 		}
 
 		if(wan_bonding_support && orig_bond_wan == "1"){
-			if(document.form.switch_stb_x.value == "4" || document.form.switch_stb_x.value == "6"){
-				var msg = "<#WANAggregation_PortConflict_hint1#>";
+			if(wanAggr_p2_conflicts_w_stb_port(document.form.switch_stb_x.value, wanAggr_p2_num(orig_wnaports_bond))){
+				var msg = "<#WANAggregation_PortConflict_hint1#>".replace(/LAN-*\D* 4/, wanAggr_p2_name(orig_wnaports_bond));
 				if(confirm(msg)){
 					document.form.bond_wan.disabled = false;
 					document.form.bond_wan.value = "0";
@@ -496,6 +541,7 @@ function applyRule(){
 			}
 		}
 
+		turn_off_lacp_if_conflicts();
 		showLoading();
 		document.form.submit();
 	}
@@ -1051,7 +1097,7 @@ function change_port_settings(val, changed){
 	}
 	else if(val == "56"){
 		if(changed){
-			var msg="Change IPTV/ VOIP port settings to LAN5/ LAN6 will disable Bonding/ Link aggregation function. Are you sure to do it?";
+			var msg="<#NAT_lacp_disable_note#>";	/*Untranslated*/
 			if(lacp_enabled){
 				if(!confirm(msg)){
 					document.form.iptv_port_settings.value = "12";
@@ -1077,13 +1123,49 @@ function change_port_settings(val, changed){
 
 function show_gaming_note(val){
 	if(val == "12")
-		document.getElementById("gaming_note").innerHTML = "Gaming Ports are set up in LAN1 and LAN2. If you would like to use Gaming Ports, please choose LAN 5/ LAN 6 for your IPTV or VoIP port.";//untranslated
+		document.getElementById("gaming_note").innerHTML = "<#RouterConfig_GW_GamingPortsNote_12#>";
 	else if(val == "56")
-		document.getElementById("gaming_note").innerHTML = "Link aggregation is configured in LAN 5 and LAN6. If you would like to use link aggregation, please choose LAN 1/ LAN 2 for your IPTV or VoIP port.";//untranslated
+		document.getElementById("gaming_note").innerHTML = "<#RouterConfig_GW_GamingPortsNote_56#>";
 	document.getElementById("gaming_note_div").style.display = "";
 }
 
-function change_mr_enable(switch_stb_x){
+function create_mr_select(val)
+{
+	var select = document.form.mr_enable_x;
+
+	if(improxy_support){
+		add_option(select, "<#WLANConfig11b_WirelessCtrl_button1name#> IGMP", "1", 0);
+	//	add_option(select, "<#WLANConfig11b_WirelessCtrl_button1name#> MLD", "2", 0);
+		add_option(select, "<#WLANConfig11b_WirelessCtrl_button1name#> IGMP & MLD", "3", 0);
+	}
+	else
+		add_option(select, "<#WLANConfig11b_WirelessCtrl_button1name#>", "1", 0);
+
+	select.value = val;
+	if(select.selectedIndex < 0)
+		select.selectedIndex = 0;
+
+	change_mr_enable(select.value);
+}
+
+function change_mr_enable(val){
+	var igmp_enable = (val == "1" || val == "3") && (improxy_support || hnd_support);
+	var mld_enable = (val == "2" || val == "3") && improxy_support;
+	var qleave_enable = val && (val != "0");
+	inputCtrl(document.form.mr_igmp_ver, igmp_enable ? 1 : 0);
+	inputCtrl(document.form.mr_mld_ver, mld_enable ? 1 : 0);
+	inputCtrl(document.form.mr_qleave_x, qleave_enable ? 1 : 0);
+}
+
+function update_mr_mswan_idx(){
+	for(var i = 1; i < MSWAN_List_Pri.length; i++){
+		if (MSWAN_List_Pri[i][0] == "1") {
+			add_option(document.form.mr_mswan_idx, i, i, (mr_mswan_idx_orig==i)?1:0);
+		}
+	}
+}
+
+function change_switch_stb(switch_stb_x){
 	if(hnd_support || based_modelid == "BLUECAVE"){
 		if(switch_stb_x != "0"){
 			document.getElementById("mr_enable_x").style.display = "none";
@@ -1094,15 +1176,19 @@ function change_mr_enable(switch_stb_x){
 			document.getElementById("mr_enable_x").style.display = "";
 			document.getElementById("mr_disable").style.display = "none";
 			document.form.mr_enable_x.value = orig_mr_enable;
+			if(document.form.mr_enable_x.selectedIndex < 0)
+				document.form.mr_enable_x.selectedIndex = 0;
 		}
 	}
 	else
 		document.getElementById("mr_disable").style.display = "none";
+
+	change_mr_enable(document.form.mr_enable_x.value);
 }
 </script>
 </head>
 
-<body onload="initial();" onunLoad="return unload_body();">
+<body onload="initial();" onunLoad="return unload_body();" class="bg">
 <div id="TopBanner"></div>
 <div id="hiddenMask" class="popup_bg">
 	<table cellpadding="5" cellspacing="0" id="dr_sweet_advise" class="dr_sweet_advise" align="center">
@@ -1416,10 +1502,7 @@ function change_mr_enable(switch_stb_x){
 			<div id="IPTV_desc_DualWAN_BRTAC828" class="formfontdesc" style="display:none;">
 				<#LANHostConfig_displayIPTV_sectiondescBRTAC828#>
 			</div>
-	  
-	  <!-- IPTV & VoIP Setting -->
-	  
-		<!--###HTML_PREP_START###-->
+
 	  <table width="100%" border="1" align="center" cellpadding="4" cellspacing="0" bordercolor="#6b8fa3"  class="FormTable">
 	  	<thead>
 			<tr>
@@ -1427,7 +1510,7 @@ function change_mr_enable(switch_stb_x){
 			</tr>
 		</thead>
 			<tr id="port_settings" style="display:none;">
-				<th width="30%"><a class="hintstyle" href="javascript:void(0);" onClick="openHint(7,28);">IPTV/ VoIP Port Settings</a></th>
+				<th width="30%"><a class="hintstyle" href="javascript:void(0);" onClick="openHint(7,28);"><#RouterConfig_GW_LANPort_itemname#></a></th>
 				<td>
 					<select name="iptv_port_settings" class="input_option" onChange="change_port_settings(this.value, 1);" disabled>
 						<option value="12" <% nvram_match( "iptv_port_settings", "12", "selected"); %>>LAN1/ LAN2</option>
@@ -1436,7 +1519,7 @@ function change_mr_enable(switch_stb_x){
 				<div id="gaming_note_div" style="display: none;"><span id="gaming_note"></span><div>
 			</td>
 			</tr>
-	    	<tr>
+		<tr id="isp_profile_tr">
 		    	<th width="30%"><a class="hintstyle" href="javascript:void(0);" onClick="openHint(7,28);"><#Select_ISPfile#></a></th>
 			<td>
 				<select id="switch_wantag" name="switch_wantag" class="input_option" onChange="ISP_Profile_Selection(this.value)">
@@ -1446,8 +1529,9 @@ function change_mr_enable(switch_stb_x){
 		<tr id="wan_stb_x">
 		<th width="30%"><#Layer3Forwarding_x_STB_itemname#></th>
 		<td align="left">
-		    <select id="switch_stb_x0" name="switch_stb_x0" class="input_option" onchange="control_wans_primary(this.value);change_mr_enable(this.value);">
+		    <select id="switch_stb_x0" name="switch_stb_x0" class="input_option" onchange="control_wans_primary(this.value);change_switch_stb(this.value);">
 		    </select>
+			<span id="dsl_vlan_check" style="color:#FFFFFF; display:none;"><input type="checkbox" name="dslx_rmvlan_check" id="dslx_rmvlan_check" value="" onClick="change_rmvlan();"> Remove VLAN TAG from DSL WAN</input></span>
 		</td>
 		</tr>
 		<tr id="tr_wans_primary" style="display:none;">
@@ -1493,60 +1577,7 @@ function change_mr_enable(switch_stb_x){
 			</td>
 		</tr>
 		</table>
-<!--###HTML_PREP_ELSE###-->
-<!--
-[DSL-N55U][DSL-N55U-B]
-{DSL do not support unifw}
-	  <table width="100%" border="1" align="center" cellpadding="4" cellspacing="0" bordercolor="#6b8fa3"  class="FormTable">
-	  	<thead>
-		<tr>
-            	<td colspan="2">Port</td>
-            	</tr>
-		</thead>
-		<tr id="wan_stb_x">
-		<th width="30%"><#Layer3Forwarding_x_STB_itemname#></th>
-		<td align="left">
-		    <select name="switch_stb_x" class="input_option">
-			<option value="0" <% nvram_match( "switch_stb_x", "0", "selected"); %>><#wl_securitylevel_0#></option>
-			<option value="1" <% nvram_match( "switch_stb_x", "1", "selected"); %>>LAN1</option>
-			<option value="2" <% nvram_match( "switch_stb_x", "2", "selected"); %>>LAN2</option>
-			<option value="3" <% nvram_match( "switch_stb_x", "3", "selected"); %>>LAN3</option>
-			<option value="4" <% nvram_match( "switch_stb_x", "4", "selected"); %>>LAN4</option>
-			<option value="5" <% nvram_match( "switch_stb_x", "5", "selected"); %>>LAN1 & LAN2</option>
-			<option value="6" <% nvram_match( "switch_stb_x", "6", "selected"); %>>LAN3 & LAN4</option>
-		    </select>
-		</td>
-		</tr>
-		</table>
-[DSL-AC68U]
-	<table width="100%" border="1" align="center" cellpadding="4" cellspacing="0" bordercolor="#6b8fa3"  class="FormTable">
-		<thead>
-			<tr>
-				<td colspan="2">Port</td>
-			</tr>
-		</thead>
-		<tr id="wan_stb_x">
-			<th width="30%"><#Layer3Forwarding_x_STB_itemname#></th>
-			<td align="left">
-				<select name="switch_stb_x" class="input_option">
-				<option value="0" <% nvram_match( "switch_stb_x", "0", "selected"); %>><#wl_securitylevel_0#></option>
-				<option value="1" <% nvram_match( "switch_stb_x", "1", "selected"); %>>LAN1</option>
-				<option value="2" <% nvram_match( "switch_stb_x", "2", "selected"); %>>LAN2</option>
-				<option value="3" <% nvram_match( "switch_stb_x", "3", "selected"); %>>LAN3</option>
-				<option value="4" <% nvram_match( "switch_stb_x", "4", "selected"); %>>LAN4</option>
-				<option value="5" <% nvram_match( "switch_stb_x", "5", "selected"); %>>LAN1 & LAN2</option>
-				<option value="6" <% nvram_match( "switch_stb_x", "6", "selected"); %>>LAN3 & LAN4</option>
-				</select>
-				<input type="checkbox" name="dslx_rmvlan_check" id="dslx_rmvlan_check" value="" onClick="change_rmvlan();"> Remove VLAN TAG from DSL WAN</input>
-			</td>
-		</tr>
-	</table>
--->
-<!--###HTML_PREP_END###-->	  
-	  
 
-		<!-- End of IPTV & VoIP -->	  
-	  
 		  <table width="100%" border="1" align="center" cellpadding="4" cellspacing="0" bordercolor="#6b8fa3"  class="FormTable" style="margin-top:10px;">
 	  	<thead>
 		<tr>
@@ -1565,21 +1596,56 @@ function change_mr_enable(switch_stb_x){
 				</select>
 			</td>
 		</tr>
-		<tr id="mr_enable_field" style="display:none;">
-			<th><a class="hintstyle" href="javascript:void(0);" onClick="openHint(5,11);"><#RouterConfig_GWMulticastEnable_itemname#> (IGMP Proxy)</a></th>
+		<tr style="display:none;">
+			<th>WAN index</th>
 			<td>
-				<select id="mr_enable_x" name="mr_enable_x" class="input_option">
+				<select name="mr_mswan_idx" class="input_option">
+					<option value="0" <% nvram_match("mr_mswan_idx", "0","selected"); %> >0</option>
+				</select>
+			</td>
+		</tr>
+		<tr id="mr_enable_field" style="display:none;">
+			<th><a class="hintstyle" href="javascript:void(0);" onClick="openHint(5,11);"><#RouterConfig_GWMulticastEnable_itemname#></a></th>
+			<td>
+				<select id="mr_enable_x" name="mr_enable_x" class="input_option" onChange="change_mr_enable(this.value);">
 					<option value="0" <% nvram_match("mr_enable_x", "0","selected"); %> ><#WLANConfig11b_WirelessCtrl_buttonname#></option>
-					<option value="1" <% nvram_match("mr_enable_x", "1","selected"); %> ><#WLANConfig11b_WirelessCtrl_button1name#></option>
+				<!--	<option value="1" <% nvram_match("mr_enable_x", "1","selected"); %> ><#WLANConfig11b_WirelessCtrl_button1name#></option> -->
 				</select>
 				<span id="mr_hint" style="display:none;">( <#RouterConfig_GWMulticastEnable_hint#> )</span>
 				<div id="mr_disable" style="display:none;">
 					<span style="color:#FFF;"><#WLANConfig11b_WirelessCtrl_buttonname#></span>
-					<span style="margin-left: 5px;">(Due to hardware limitation, IGMP proxy cannot co-exist with IPTV function.)</span><!--untranslated-->
+					<span style="margin-left: 5px;"><#RouterConfig_GWMulticastEnable_hint2#></span>
 				</div>
 			</td>
 		</tr>
-
+		<tr style="display:none;">
+			<th><a class="hintstyle" href="javascript:void(0);" onClick="openHint(5,14);"><#RouterConfig_IGMPver_itemname#></a></th>
+			<td>
+				<select name="mr_igmp_ver" class="input_option">
+					<option value="1" <% nvram_match("mr_igmp_ver", "1","selected"); %> >IGMP v1</option>
+					<option value="2" <% nvram_match("mr_igmp_ver", "2","selected"); %> >IGMP v2</option>
+					<option value="3" <% nvram_match("mr_igmp_ver", "3","selected"); %> >IGMP v3</option>
+				</select>
+			</td>
+		</tr>
+		<tr style="display:none;">
+			<th><a class="hintstyle" href="javascript:void(0);" onClick="openHint(5,15);"><#RouterConfig_MLDver_itemname#></a></th>
+			<td>
+				<select name="mr_mld_ver" class="input_option">
+					<option value="1" <% nvram_match("mr_mld_ver", "1","selected"); %> >MLD v1</option>
+					<option value="2" <% nvram_match("mr_mld_ver", "2","selected"); %> >MLD v2</option>
+				</select>
+			</td>
+		</tr>
+		<tr style="display:none;">
+			<th><a class="hintstyle" href="javascript:void(0);" onClick="openHint(5,16);">Enable Fast Leave</a></th>
+			<td>
+				<select name="mr_qleave_x" class="input_option">
+					<option value="0" <% nvram_match("mr_qleave_x", "0","selected"); %> ><#WLANConfig11b_WirelessCtrl_buttonname#></option>
+					<option value="1" <% nvram_match("mr_qleave_x", "1","selected"); %> ><#WLANConfig11b_WirelessCtrl_button1name#></option>
+				</select>
+			</td>
+		</tr>
 		<tr id="enable_eff_multicast_forward" style="display:none;">
 			<th><#WLANConfig11b_x_Emf_itemname#></th>
 			<td>
@@ -1594,15 +1660,6 @@ function change_mr_enable(switch_stb_x){
 			<th><a class="hintstyle" href="javascript:void(0);" onClick="openHint(6, 6);"><#RouterConfig_IPTV_itemname#></a></th>
 			<td>
 				<input id="udpxy_enable_x" type="text" maxlength="5" class="input_6_table" name="udpxy_enable_x" value="<% nvram_get("udpxy_enable_x"); %>" onkeypress="return validator.isNumber(this,event);" autocorrect="off" autocapitalize="off">
-			</td>
-		</tr>
-		<tr id="igmp_default_ver_field" style="display:none;">
-			<th>IGMP default version</th>
-			<td>
-				<select name="igmp_default_version" class="input_option">
-					<option value="2" <% nvram_match("igmp_default_version", "2","selected"); %> >2</option>
-					<option value="3" <% nvram_match("igmp_default_version", "3","selected"); %> >3</option>
-				</select>
 			</td>
 		</tr>
 		</table>
